+++ modules/AWSPowerShell/AWSPowerShellCompleters.psm1
@@ -36374,6 +36374,7 @@ $OUTP_SelectMap = @{
                "Get-OUTPOutpostInstanceType",
                "Get-OUTPSite",
                "Get-OUTPSiteAddress",
+               "Get-OUTPAssetList",
                "Get-OUTPCatalogItemList",
                "Get-OUTPOrderList",
                "Get-OUTPOutpostList",