+++ modules/AWSPowerShell/Cmdlets/Outposts/Basic/Get-OUTPSiteList-Cmdlet.cs
@@ -28,9 +28,9 @@ using Amazon.Outposts.Model;
 namespace Amazon.PowerShell.Cmdlets.OUTP
 {
     /// <summary>
-    /// Create a list of the Outpost sites for your Amazon Web Services account. Add operating
-    /// address filters to your request to return a more specific list of results. Use filters
-    /// to match site city, country code, or state/region of the operating address. 
+    /// Lists the Outpost sites for your Amazon Web Services account. Add operating address
+    /// filters to your request to return a more specific list of results. Use filters to
+    /// match site city, country code, or state/region of the operating address. 
     /// 
     ///  
     /// <para>